// src/generated/zod/inputTypeSchemas/DonorInfoMaxOrderByAggregateInputSchema.ts
import type { Prisma } from '*/generated/prisma'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const DonorInfoMaxOrderByAggregateInputSchema: z.ZodType<Prisma.DonorInfoMaxOrderByAggregateInput> =
	z
		.object({
			id: z.lazy(() => SortOrderSchema).optional(),
			name: z.lazy(() => SortOrderSchema).optional(),
			avatar: z.lazy(() => SortOrderSchema).optional(),
			description: z.lazy(() => SortOrderSchema).optional(),
			donationAmount: z.lazy(() => SortOrderSchema).optional(),
			donationDate: z.lazy(() => SortOrderSchema).optional(),
			operatorId: z.lazy(() => SortOrderSchema).optional(),
			createdAt: z.lazy(() => SortOrderSchema).optional(),
			updatedAt: z.lazy(() => SortOrderSchema).optional(),
		})
		.strict()

export default DonorInfoMaxOrderByAggregateInputSchema